# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Marketplace API for the Agri Supply Chain Platform
 * OpenAPI spec version: 0.1.0
 */
import type { QualityGrade } from './qualityGrade';

export interface Listing {
  id: string;
  sellerId: string;
  cropType: string;
  quantity: number;
  availableQuantity: number;
  unit: string;
  pricePerUnit: number;
  location: string;
  /** @nullable */
  photoUrl?: string | null;
  qualityGrade?: QualityGrade | null;
  /** @nullable */
  qualityReason?: string | null;
  createdAt: Date;
  updatedAt: Date;
}