# lib/api-zod/src/generated/types/qualityGrade.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Marketplace API for the Agri Supply Chain Platform
 * OpenAPI spec version: 0.1.0
 */

export type QualityGrade = typeof QualityGrade[keyof typeof QualityGrade];


export const QualityGrade = {
  Good: 'Good',
  Medium: 'Medium',
  Poor: 'Poor',
} as const;