# lib/api-zod/src/generated/types/storageUploadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Marketplace API for the Agri Supply Chain Platform
 * OpenAPI spec version: 0.1.0
 */
import type { StorageUploadInputContentType } from './storageUploadInputContentType';

export interface StorageUploadInput {
  sellerId: string;
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 8388608
     */
  size: number;
  contentType: StorageUploadInputContentType;
}