# lib/api-zod/src/generated/types/storageUploadResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Marketplace API for the Agri Supply Chain Platform
 * OpenAPI spec version: 0.1.0
 */

export interface StorageUploadResponse {
  uploadURL: string;
  objectPath: string;
}